+++ src/components/select/simple-select/simple-select.stories.tsx
@@ -15,7 +15,7 @@ import Box from "../../box";
 import Typography from "../../typography";
 import generateStyledSystemProps from "../../../../.storybook/utils/styled-system-props";
 
-import SimpleSelect from "./simple-select.component";
+import SimpleSelect, { SimpleSelectProps } from "./simple-select.component";
 
 const styledSystemProps = generateStyledSystemProps({
   margin: true,
@@ -854,3 +854,47 @@ export const SelectWithDynamicallyAddedOption: Story = () => {
   );
 };
 SelectWithDynamicallyAddedOption.storyName = "Dynamically Adding Options";
+
+export const ListWidth: Story = () => {
+  const [listPlacement, setListPlacement] = useState<
+    SimpleSelectProps["listPlacement"]
+  >("bottom-end");
+  const [value, setValue] = useState("");
+  const handleChange = (ev: React.ChangeEvent<HTMLInputElement>) => {
+    setValue(ev.target.value);
+  };
+  return (
+    <>
+      <Button mr={1} onClick={() => setListPlacement("top-end")}>
+        Top end
+      </Button>
+      <Button mr={1} onClick={() => setListPlacement("bottom-end")}>
+        Bottom end
+      </Button>
+      <Button mr={1} onClick={() => setListPlacement("top-start")}>
+        Top start
+      </Button>
+      <Button onClick={() => setListPlacement("bottom-start")}>
+        Bottom Start
+      </Button>
+      <Box mt="200px" ml="200px" width="200px">
+        <Select
+          name="listWidth"
+          id="listWidth"
+          label="color"
+          labelInline
+          listWidth={350}
+          listPlacement={listPlacement}
+          value={value}
+          onChange={handleChange}
+        >
+          <Option text="Amber" value="1" />
+          <Option text="Black" value="2" />
+          <Option text="Blue" value="3" />
+        </Select>
+      </Box>
+    </>
+  );
+};
+ListWidth.storyName = "List width";
+ListWidth.parameters = { chromatic: { disableSnapshot: true } };